requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

# capstoneproject/archive/pdf-pre-processing.ipynb: 22,23
# capstoneproject/predictions.py: 6
Keras == 2.4.3

# capstoneproject/archive/pdf-pre-processing.ipynb: 1
PyPDF2 == 1.26.0

# capstoneproject/.ipynb_checkpoints/historical_data_collection_1-6-checkpoint.ipynb: 17
# capstoneproject/.ipynb_checkpoints/live_data_collection_2-6-checkpoint.ipynb: 17
# capstoneproject/archive/pdf-pre-processing.ipynb: 2
# capstoneproject/historical_data_collection_1-6.ipynb: 17
# capstoneproject/live_data_collection_2-6.ipynb: 17
# capstoneproject/main-workflow_6-6.py: 24
gensim == 3.8.3

# capstoneproject/historical.py: 3
# capstoneproject/main-workflow_6-6.py: 27
# capstoneproject/predictions.py: 3
# capstoneproject/top_mentions.py: 3
matplotlib == 3.3.2

# capstoneproject/.ipynb_checkpoints/historical_data_collection_1-6-checkpoint.ipynb: 10,11,12,13,14
# capstoneproject/.ipynb_checkpoints/live_data_collection_2-6-checkpoint.ipynb: 10,11,12,13,14
# capstoneproject/archive/pdf-pre-processing.ipynb: 19,20,21,39
# capstoneproject/historical_data_collection_1-6.ipynb: 10,11,12,13,14
# capstoneproject/live_data_collection_2-6.ipynb: 10,11,12,13,14
# capstoneproject/main-workflow_6-6.py: 14,15,16,17,18
nltk == 3.5

# capstoneproject/.ipynb_checkpoints/historical_data_collection_1-6-checkpoint.ipynb: 21
# capstoneproject/.ipynb_checkpoints/live_data_collection_2-6-checkpoint.ipynb: 22
# capstoneproject/historical_data_collection_1-6.ipynb: 21
# capstoneproject/live_data_collection_2-6.ipynb: 22
# capstoneproject/main-workflow_6-6.py: 33
numpy == 1.19.2

# capstoneproject/.ipynb_checkpoints/historical_data_collection_1-6-checkpoint.ipynb: 22
# capstoneproject/.ipynb_checkpoints/live_data_collection_2-6-checkpoint.ipynb: 23
# capstoneproject/historical.py: 2
# capstoneproject/historical_data_collection_1-6.ipynb: 22
# capstoneproject/live_data_collection_2-6.ipynb: 23
# capstoneproject/main-workflow_6-6.py: 34
# capstoneproject/predictions.py: 2
# capstoneproject/top_mentions.py: 2
# capstoneproject/tweet_search.py: 2
pandas == 1.1.5

# capstoneproject/.ipynb_checkpoints/historical_data_collection_1-6-checkpoint.ipynb: 30
# capstoneproject/.ipynb_checkpoints/live_data_collection_2-6-checkpoint.ipynb: 31
# capstoneproject/historical_data_collection_1-6.ipynb: 30
# capstoneproject/live_data_collection_2-6.ipynb: 31
# capstoneproject/main-workflow_6-6.py: 43
python_dotenv == 0.15.0

# capstoneproject/.ipynb_checkpoints/historical_data_collection_1-6-checkpoint.ipynb: 23
# capstoneproject/.ipynb_checkpoints/live_data_collection_2-6-checkpoint.ipynb: 24
# capstoneproject/historical_data_collection_1-6.ipynb: 23
# capstoneproject/live_data_collection_2-6.ipynb: 24
# capstoneproject/main-workflow_6-6.py: 35
# capstoneproject/predictions.py: 4
requests == 2.25.1

# capstoneproject/.ipynb_checkpoints/historical_data_collection_1-6-checkpoint.ipynb: 28
# capstoneproject/.ipynb_checkpoints/live_data_collection_2-6-checkpoint.ipynb: 29
# capstoneproject/historical_data_collection_1-6.ipynb: 28
# capstoneproject/live_data_collection_2-6.ipynb: 29
# capstoneproject/main-workflow_6-6.py: 41
retry == 0.9.2

# capstoneproject/predictions.py: 5
scikit_learn == 0.23.2

# capstoneproject/main-workflow_6-6.py: 28
seaborn == 0.11.1

# capstoneproject/.ipynb_checkpoints/historical_data_collection_1-6-checkpoint.ipynb: 7
# capstoneproject/.ipynb_checkpoints/live_data_collection_2-6-checkpoint.ipynb: 7
# capstoneproject/historical_data_collection_1-6.ipynb: 7
# capstoneproject/live_data_collection_2-6.ipynb: 7
# capstoneproject/main-workflow_6-6.py: 11
snscrape == 0.3.5.dev95+g5cd3b7d

# capstoneproject/main-workflow_6-6.py: 47
statsmodels == 0.12.2

# capstoneproject/historical.py: 1
# capstoneproject/main-workflow_6-6.py: 21
# capstoneproject/main.py: 6
# capstoneproject/predictions.py: 1
# capstoneproject/top_mentions.py: 1
# capstoneproject/tweet_search.py: 1
streamlit == 0.78.0

# capstoneproject/predictions.py: 7
tensorflow == 2.4.1

# capstoneproject/.ipynb_checkpoints/historical_data_collection_1-6-checkpoint.ipynb: 4
# capstoneproject/.ipynb_checkpoints/live_data_collection_2-6-checkpoint.ipynb: 4
# capstoneproject/historical_data_collection_1-6.ipynb: 4
# capstoneproject/live_data_collection_2-6.ipynb: 4
# capstoneproject/main-workflow_6-6.py: 8
tweepy == 3.10.0
